Script which removes DEP components (DEPNotify, LaunchDaemon etc..)
NOTE :  This script does not remove the JSS record!

Script which removes local DEP components (DEPNotify, LaunchDaemon etc..). Can be used in the event that there is a failure with the DEP process. It should remove all DEP components to ready the machine for the QuickAdd package to be run. Also requests if the jamf framework is to be removed and runs the "remove-NoLoAD.sh" script to revert login preferences back to default.