Added logic to deal with computer name problem
In some cases EdLAN DB was returning "error" as the machine name. The code originally only checked for any return that wasn't null, assuming that any name was a valid return. Added login so that if "error" is returned then it will move on to using a "dig" to obtain DNS Name.